Refactor RerunHelper to reduce API calls
